Restructured the views directory in the data upload app
