+++ timetable_solutions/interfaces/data_upload/views/__init__.py
@@ -1,0 +1,4 @@
+# Upload views
+from .upload_view_base_class import UploadPage
+from .upload_views import (TeacherListUpload, PupilListUpload, ClassroomListUpload, TimetableStructureUpload,
+                           UnsolvedClassUpload, FixedClassUpload)
